dev-requirements.txt
django-stubs
ruff
